Refactored models.py
