+++ netbox_dns/models/__init__.py
@@ -1,0 +1,4 @@
+from .zone import *
+from .nameserver import *
+from .record import *
+from .view import *
